q#67: Display correct output when multiple calls are on a line
Fixes #67. When there are multiple calls on a line, we need to
correctly identify what call we are showing. We can accomplish
this by exploiting the fact we can get the bytecode and the
currently executing instruction from the caller's frame.

Our goal is to find the "call position" of the instruction
within the line. For instance, in the code below:

```python
a(b(), c())
```

the b() call has call position 0, the c() call has call position 1,
and the a() call has call position 2. (In short, call position is
when the call will get evaluated).

We can find this by counting the number of CALL instruction between
the currently executing instruction and the instruction that starts
the line. Once we have the call position, we can give it to
a NodeVisitor subclass, which can use that infomation to find the
corresponding CALL node.

Additionally, we can no longer use the end of the line as the final
offset (since there can be multiple calls on a line), so we use
the call node's end_col_offset instead.